// frontend/src/redux/actions/action-types.ts
//Resource parameterType actions

export enum ResourceTypeAsync {
  GET_ALL = 'TYPE_REQUEST_ALL',
  GET_ALL_SUCCESS = 'TYPE_REQUEST_ALL_SUCCESS',
  GET_ALL_FAILURE = 'TYPE_REQUEST_ALL_FAILURE',
  GET = 'TYPE_GET_REQUEST',
  GET_SUCCESS = 'TYPE_GET_SUCCESS',
  GET_FAILURE = 'TYPE_GET_FAILURE',
  UPDATE = 'TYPE_UPDATE',
  UPDATE_SUCCESS = 'TYPE_UPDATE_SUCCES',
  UPDATE_FAILURE = 'TYPE_UPDATE_FAILURE',
  REQUEST_ALL_ACTIVE = 'REQUEST_ALL_ACTIVE',
  REQUEST_ALL_ACTIVE_SUCCESS = 'REQUEST_ALL_ACTIVE_SUCCESS',
  REQUEST_ALL_ACTIVE_FAILURE = 'REQUEST_ALL_ACTIVE_FAILURE',
  REQUEST_SUGGESTIONS = 'TYPE_REQUEST_SUGGESTIONS',
  REQUEST_SUGGESTIONS_SUCCES = 'TYPE_REQUEST_SUGGESTIONS_SUCCESS',
  REQUEST_SUGGESTIONS_FAILURE = 'TYPE_REQUEST_SUGGESTIONS_FAILURE',
}

export enum ResourceTypesActions {
  CLEAR_SUGGESTIONS = 'TYPE_CLEAR_SUGGESTIONS',
}

export enum ParameterAsync {
  GET_ALL = 'PARAM_REQUEST_ALL',
  GET_ALL_SUCCESS = 'PARAM_REQUEST_ALL_SUCCESS',
  GET_ALL_FAILURE = 'PARAM_REQUEST_ALL_FAILURE',
  GET_ALL_FULL_PARAMETERS = 'FULL_PARAM_REQUEST_ALL',
  GET_ALL_FULL_PARAMETERS_SUCCESS = 'FULL_PARAM_REQUEST_ALL_SUCCESS',
  GET_ALL_FULL_PARAMETERS_FAILURE = 'FULL_PARAM_REQUEST_ALL_FAILURE',
}

export enum ParameterValuesAsync {
  GET_PARAMETER_VALUE = 'GET_PARAMETER_VALUE',
  GET_PARAMETER_VALUE_SUCCESS = 'GET_PARAMETER_VALUE_SUCCESS',
  GET_PARAMETER_VALUE_FAILURE = 'GET_PARAMETER_VALUE_FAILURE',
  PARAMETER_VALUE_CREATE = 'PARAMETER_VALUE_CREATE',
  PARAMETER_VALUE_CREATE_SUCCESS = 'PARAMETER_VALUE_CREATE_SUCCESS',
  PARAMETER_VALUE_CREATE_FAILURE = 'PARAMETER_VALUE_CREATE_FAILURE',
  PARAMETER_VALUE_UPDATE = 'PARAMETER_VALUE_UPDATE',
  PARAMETER_VALUE_UPDATE_SUCCESS = 'PARAMETER_VALUE_UPDATE_SUCCESS',
  PARAMETER_VALUE_UPDATE_FAILURE = 'PARAMETER_VALUE_UPDATE_FAILURE',
}

export enum EmployeeAsync {
  EMPLOYEE_ALL = 'EMPLOYEE_ALL',
  EMPLOYEE_ALL_SUCCESS = 'EMPLOYEE_ALL_SUCCESS',
  EMPLOYEE_ALL_FAILURE = 'EMPLOYEE_ALL_FAILURE',
  FULL_EMPLOYEES_REQUEST = 'FULL_EMPLOYEES_REQUEST',
  FULL_EMPLOYEES_SUCCESS = 'FULL_EMPLOYEES_SUCCESS',
  FULL_EMPLOYEES_FAILURE = 'FULL_EMPLOYEES_FAILURE',
  GET_EMPLOYEE_REQUEST = 'GET_EMPLOYEE_REQUEST',
  GET_EMPLOYEE_SUCCESS = 'GET_EMPLOYEE_SUCCESS',
  GET_EMPLOYEE_FAILURE = 'GET_EMPLOYEE_FAILURE',
  CHANGE_ROLE_REQUEST = 'CHANGE_ROLE_REQUEST',
  CHANGE_ROLE_SUCCESS = 'CHANGE_ROLE_SUCCESS',
  CHANGE_ROLE_FAILURE = 'CHANGE_ROLE_FAILURE',
}

export enum PoolAsync {
  POOL_CREATE = 'POOL_CREATE',
  POOL_CREATE_SUCCESS = 'POOL_CREATE_SUCCESS',
  POOL_CREATE_FAILURE = 'POOL_CREATE_FAILURE',
  GET = 'POOL_REQUEST',
  GET_SUCCESS = 'POOL_GET_SUCCESS',
  GET_FAILURE = 'POOL_GET_FAILURE',
  POOL_UPDATE = 'POOL_UPDATE',
  POOL_UPDATE_SUCCESS = 'POOL_UPDATE_SUCCESS',
  POOL_UPDATE_FAILURE = 'POOL_UPDATE_FAILURE',
}

export enum RequestAsync {
  GET_ALL = 'REQUEST_REQUEST_ALL',
  GET_ALL_SUCCESS = 'REQUEST_REQUEST_ALL_SUCCESS',
  GET_ALL_FAILURE = 'REQUEST_REQUEST_ALL_FAILURE',
}

export enum ProjectAsync {
  GET_SUGGESTIONS = 'PROJECT_GET_SUGGESTIONS',
  GET_SUGGESTIONS_SUCCESS = 'PROJECT_GET_SUGGESTIONS_SUCCESS',
  GET_SUGGESTIONS_FAILURE = 'PROJECT_GET_SUGGESTIONS_FAILURE',
}

export enum ProjectActions {
  CLEAR_SUGGESTIONS = 'PROJECT_CLEAR_SUGGESTIONS',
}

export enum ResourceTypeCreateAsync {
  TYPE_CREATE = 'TYPE_CREATE',
  TYPE_CREATE_SUCCESS = 'TYPE_CREATE_SUCCESS',
  TYPE_CREATE_FAILURE = 'TYPE_CREATE_FAILURE',
  GET_ALL = 'TYPE_REQUEST',
  GET_SUCCESS = 'TYPE_GET_SUCCESS',
  GET_FAILURE = 'TYPE_GET_FAILURE',
  CHECK_NAME = 'CHECK_NAME',
  CHECK_NAME_SUCCESS = 'CHECK_NAME_SUCCESS',
  CHECK_NAME_FAILURE = 'CHECK_NAME_FAILURE',
}

export enum SettingsAsync {
  GET_SETTINGS_BY_GROUP = 'GET_SETTINGS_BY_GROUP',
  GET_SETTINGS_BY_GROUP_SUCCESS = 'GET_SETTINGS_BY_GROUP_SUCCESS',
  GET_SETTINGS_BY_GROUP_FAILURE = 'GET_SETTINGS_BY_GROUP_FAILURE',
  SETTINGS_SAVE = 'SETTINGS_SAVE',
  SETTINGS_SAVE_SUCCESS = 'SETTINGS_SAVE_SUCCESS',
  SETTINGS_SAVE_FAILURE = 'SETTINGS_SAVE_FAILURE',
}